Move Observability to pkg (#845)
* small docs typo

* Move observability to pkg

* Run goimports -w .
