+++ pkg/observability/generic.go
