+++ pkg/observability/resource.go
